got complete alla tasks button to work
