+++ code/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^1.5.1",
     "babel-eslint": "^10.0.3",
+    "dayjs": "^1.10.4",
     "eslint": "^6.7.1",
     "eslint-config-airbnb": "^18.0.1",
     "eslint-plugin-import": "^2.18.2",
@@ -12,10 +13,12 @@
     "eslint-plugin-react": "^7.16.0",
     "eslint-plugin-react-hooks": "^2.3.0",
     "react": "^16.12.0",
+    "react-dayjs": "^0.3.2",
     "react-dom": "^16.12.0",
     "react-redux": "^7.2.3",
     "react-scripts": "3.2.0",
-    "styled-components": "^5.2.3"
+    "styled-components": "^5.2.3",
+    "uniqid": "^5.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
