+++ code/src/components/TaskSummary.js
@@ -1,24 +1,15 @@
-import React, { useState} from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-
+import React from 'react';
 import styled from 'styled-components';
 
-import tasks from '../reducers/tasks';
+import UnfinishedTasksCounter from './UnfinishedTasksCounter';
+import CompleteAllTasksButton from './CompleteAllTasksButton';
 
 const TaskSummary = () => {
-  const allTasks = useSelector(store => store.tasks.items)
-  const dispatch = useDispatch();
-  
-  const onClear = () => {
-    dispatch(tasks.actions.setAllTaskstoCompleted());
-  };
-
-  const uncompleteTasks = allTasks.filter(item => item.isComplete===false).length;
 
   return (
     <div>
-      <p>{uncompleteTasks} unfinished tasks</p>
-      <button onClick={onClear}>I have done it all!</button>
+      <UnfinishedTasksCounter/>
+      <CompleteAllTasksButton/>
     </div> 
   )
 }
